test: Elect repro-359 to shared e2e test
It failed the app router debounce queue reset locally,
so making sure nothing else breaks elsewhere.
