+++ packages/e2e/next/src/app/app/(shared)/repro-359/page.tsx
@@ -1,0 +1,10 @@
+import { Repro359 } from 'e2e-shared/specs/repro-359'
+import { Suspense } from 'react'
+
+export default function Page() {
+  return (
+    <Suspense>
+      <Repro359 />
+    </Suspense>
+  )
+}
