+++ packages/e2e/react/src/routes/repro-359.tsx
@@ -1,0 +1,3 @@
+import { Repro359 } from 'e2e-shared/specs/repro-359'
+
+export default Repro359
