+++ packages/e2e/remix/app/routes/repro-359.tsx
@@ -1,0 +1,3 @@
+import { Repro359 } from 'e2e-shared/specs/repro-359'
+
+export default Repro359
