+++ packages/e2e/remix/cypress/e2e/shared/repro-359.cy.ts
@@ -1,0 +1,5 @@
+import { testRepro359 } from 'e2e-shared/specs/repro-359.cy'
+
+testRepro359({
+  path: '/repro-359'
+})
